API method to find movie commentary/reviews
I have stored the YouTube API key in a local .env file that is referenced in my config.py. This may change if requested by the professor.

I have created youtube.py, which has the method search_movie_reviews that outputs a list of 10 movie reviews and commentaries about that movie.
This was done using the youtube.search().list() method with results appended to a list videos as a return value.

I used youtube.py to edit main.py to create a new API method that returns movie reviews for a specified movie name

I updated test_main.py and created test_youtube.py to test the new methods.
I used mocker to ensure that pytest is not using the actual API search as to not fill the quota.

I have updated the README.md to specify to the user that, at least for now, will need to provide their own API key.

I uploaded the YouTube API key to GitHub Actions Secrets and updated the test.yml accordingly.