Get a full description of a movie review
In youtube.py: Created a ‘get_video_description’ method that takes a youtube video id as a parameter and returns the full description of the video.

In main.py: Created an API method that uses the ‘get_video_description’ method to output the complete description of the video. It is available in the SwaggerUI.

In test_youtube.py: Created a test method ‘test_get_video_description’ that gets the video description of a valid video (currently with the id ‘U3dE0sYZqvI’). The test passes if the method does not return the response 'Description not found.', which would indicate that the video does not exist. Also created a test method that tests if a video does not exist. It passes if it returns the response 'Description not found.'.

In test_main.py: Created a test method ‘test_video_description’ that gets the video description of a valid video. If it does not return the JSON response {'description': 'Description not found.'}, the test passes. Also created a test method that tests if a video does not exist. It passes if it returns the JSON response {'description': 'Description not found.'}.

Issue #3